+++ tests/testthat/test-copy_to.src_presto.R
@@ -4,9 +4,7 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-context("copy_to.src_presto and db_copy_to")
-
-source("utilities.R")
+context(paste(Sys.getenv("PRESTO_TYPE", "Presto"), "copy_to.src_presto and db_copy_to"))
 
 .test_src <- function(src, test_table_name) {
   if (inherits(src, "src_presto")) {
@@ -47,24 +45,40 @@ source("utilities.R")
   )
   expect_true(dbExistsTable(con, test_table_name))
   expect_equal_data_frame(collect(tbl), test_df)
+  # a different schema works
+  DBI::dbExecute(con, paste0("CREATE SCHEMA IF NOT EXISTS testing"))
+  test_table_name_4 <- dbplyr::in_schema("testing", test_table_name)
+  tbl <- copy_to(dest = src, df = test_df, name = test_table_name_4)
+  expect_true(dbExistsTable(con, test_table_name_4))
+  expect_equal_data_frame(collect(tbl), test_df)
+  # overwriting a different schema works
+  expect_message(
+    tbl <- copy_to(
+      dest = src, df = test_df, name = test_table_name_4, overwrite = TRUE
+    ),
+    "The table .* is overwritten"
+  )
+  expect_true(dbExistsTable(con, test_table_name_4))
+  expect_equal_data_frame(collect(tbl), test_df)
+  expect_true(dbRemoveTable(con, test_table_name_4))
 }
 
 test_that("dplyr::copy_to works for src_presto", {
   src <- src_presto(
-    con = presto_default(output.timezone = "America/Los_Angeles")
+    con = setup_live_connection(output.timezone = "America/Los_Angeles")
   )
   test_table_name <- "test_copyto_srcpresto"
   .test_src(src, test_table_name)
 })
 
 test_that("dplyr::copy_to works for PrestoConnection", {
-  src <- presto_default(output.timezone = "America/Los_Angeles")
+  src <- setup_live_connection(output.timezone = "America/Los_Angeles")
   test_table_name <- "test_copyto_prestoconnection"
   .test_src(src, test_table_name)
 })
 
 test_that("dbplyr::db_copy_to works for PrestoConnection", {
-  con <- presto_default(output.timezone = "America/Los_Angeles")
+  con <- setup_live_connection(output.timezone = "America/Los_Angeles")
   test_table_name <- "test_dbcopyto"
   if (dbExistsTable(con, test_table_name)) {
     dbRemoveTable(con, test_table_name)