+++ tests/testthat/test-dbExecute.R
@@ -4,9 +4,7 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-context("dbExecute and dbGetRowsAffected")
-
-source("utilities.R")
+context(paste(Sys.getenv("PRESTO_TYPE", "Presto"), "dbExecute and dbGetRowsAffected"))
 
 test_that("dbExecute works with live database to create empty table", {
   conn <- setup_live_connection()
@@ -18,7 +16,7 @@ test_that("dbExecute works with live database to create empty table", {
     dbGetQuery(conn, paste0("DROP TABLE ", test_table_name))
   }
   expect_equal(
-    dbExecute(
+    DBI::dbExecute(
       conn,
       paste("CREATE TABLE", test_table_name, "(field1 BIGINT, field2 VARCHAR)")
     ),
@@ -41,11 +39,11 @@ test_that("dbExecute works with live database to replicate existing table", {
     dbGetQuery(conn, paste0("DROP TABLE ", test_table_name))
   }
   expect_equal(
-    dbExecute(
+    DBI::dbExecute(
       conn,
       paste("CREATE TABLE", test_table_name, "AS SELECT * FROM iris")
     ),
-    NROW(iris)
+    NROW(iris_df)
   )
   expect_true(dbExistsTable(conn, test_table_name))
   expect_equal(
@@ -65,7 +63,7 @@ test_that("dbExecute works with live database to insert values", {
   }
   old_nrow <- get_nrow(conn, test_table_name)
   expect_equal(
-    dbExecute(
+    DBI::dbExecute(
       conn,
       paste("INSERT INTO", test_table_name, "VALUES (1, 'a'), (2, 'b')")
     ),
@@ -96,7 +94,7 @@ test_that("dbExecute works with live database to rename table", {
     )
   }
   expect_equal(
-    dbExecute(
+    DBI::dbExecute(
       conn,
       paste(
         "ALTER TABLE", old_test_table_name, "RENAME TO", new_test_table_name