+++ tests/testthat/test-dbExistsTable.R
@@ -4,9 +4,7 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-context("dbExistsTable and db_has_table")
-
-source("utilities.R")
+context(paste(Sys.getenv("PRESTO_TYPE", "Presto"), "dbExistsTable and db_has_table"))
 
 test_that("dbExistsTable works with live database", {
   conn <- setup_live_connection()