+++ tests/testthat/test-dbSendQuery.R
@@ -4,9 +4,7 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-context("dbSendQuery")
-
-source("utilities.R")
+context(paste(Sys.getenv("PRESTO_TYPE", "Presto"), "dbSendQuery"))
 
 test_that("dbSendQuery works with live database", {
   conn <- setup_live_connection()