+++ tests/testthat/test-extra.credential.R
@@ -4,9 +4,7 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-context("extra.credentials")
-
-source("utilities.R")
+context(paste(Sys.getenv("PRESTO_TYPE", "Presto"), "extra.credentials"))
 
 test_that("extra.credentials should be set in the connection as a string", {
   conn <- setup_live_connection(extra.credentials = "test.token.foo=bar")