+++ tests/testthat/test-tbl.src_presto.R
@@ -6,8 +6,6 @@
 
 context("tbl.src_presto")
 
-source("utilities.R")
-
 test_that("dplyr::tbl works", {
   parts <- setup_live_dplyr_connection()
 
@@ -15,7 +13,7 @@ test_that("dplyr::tbl works", {
   iris_presto_materialized <- dplyr::collect(iris_presto, n = Inf)
   expect_that(
     nrow(iris_presto_materialized),
-    equals(nrow(iris))
+    equals(nrow(iris_df))
   )
 
   iris_from_sql <- dplyr::tbl(
@@ -25,50 +23,28 @@ test_that("dplyr::tbl works", {
   expect_equal(
     dplyr::arrange(
       iris_presto_materialized,
-      sepal_width,
-      sepal_length,
-      petal_width,
-      petal_length,
+      sepal.width,
+      sepal.length,
+      petal.width,
+      petal.length,
       species
     ),
     dplyr::arrange(
       dplyr::collect(iris_from_sql, n = Inf),
-      sepal_width,
-      sepal_length,
-      petal_width,
-      petal_length,
+      sepal.width,
+      sepal.length,
+      petal.width,
+      petal.length,
       species
     ),
   )
 })
 
 test_that("cross-schema tbl works", {
-  skip_if_not(presto_has_default())
-
-  conn <- DBI::dbConnect(
-    drv = Presto(),
-    host = "http://localhost",
-    port = 8080,
-    user = Sys.getenv("USER"),
-    catalog = "memory",
-    schema = "default"
-  )
-  DBI::dbWriteTable(conn, "iris", iris, overwrite = TRUE)
-  if (!"testing" %in% DBI::dbGetQuery(conn, "SHOW SCHEMAS")$Schema) {
-    DBI::dbExecute(conn, "CREATE SCHEMA testing")
-  }
-  conn2 <- DBI::dbConnect(
-    drv = Presto(),
-    host = "http://localhost",
-    port = 8080,
-    user = Sys.getenv("USER"),
-    catalog = "memory",
-    schema = "testing"
-  )
-  DBI::dbWriteTable(conn2, "iris2", iris, overwrite = TRUE)
+  conn <- setup_live_connection()
 
   tbl_iris <- dplyr::tbl(conn, "iris")
-  tbl_iris2 <- dplyr::tbl(conn, dbplyr::in_schema("testing", "iris2"))
+  tbl_iris2 <- dplyr::tbl(conn, dbplyr::in_schema("testing", "iris"))
   expect_equal_data_frame(
     dplyr::collect(tbl_iris),
     dplyr::collect(tbl_iris2)