+++ NEWS.md
@@ -1,3 +1,12 @@
+# RPresto 1.4.7
+
+* Fixed Trino support quirks (#254) and expanded unit tests to Trino
+* Fixed a bug whereby nested CTEs result in nested WITH. (#261)
+* Fixed `paste()` and `paste0()` translation (#266)
+* RPresto is now compatible with dbplyr 2.5.0 (#272, #274, #277)
+* `dbWriteTable(overwrite = TRUE)` makes sure that the temp table is created
+  in the same schema as the destination table (#279)
+
 # RPresto 1.4.6
 
 * `tidyr::fill()` now works with PrestoConnection. (#233, thanks to @copernican